Move training index getter to utils
