+++ snorkel/learning/__init__.py
@@ -2,7 +2,7 @@
 Subpackage for Snorkel machine learning modules.
 """
 from .constants import *
-from .disc_learning import get_train_idxs, NoiseAwareModel, TFNoiseAwareModel
+from .disc_learning import NoiseAwareModel, TFNoiseAwareModel
 from .gen_learning import (
     DEP_EXCLUSIVE,
     DEP_FIXING,
